+++ src/stac_asset/s3_client.py
@@ -1,20 +1,17 @@
 from __future__ import annotations
 
-from pathlib import Path
 from types import TracebackType
 from typing import AsyncIterator, Optional, Type
 
 import aiobotocore.session
+import botocore.config
 from aiobotocore.session import AioSession
 from botocore import UNSIGNED
-from botocore.config import Config
-from pystac import Asset
 from yarl import URL
 
 from .client import Client
-from .errors import AssetDownloadError, SchemeError
-
-DEFAULT_REGION_NAME = "us-west-2"
+from .config import DEFAULT_S3_REGION_NAME, Config
+from .errors import ContentTypeError
 
 
 class S3Client(Client):
@@ -29,19 +26,38 @@ class S3Client(Client):
     requester_pays: bool
     """If True, add `--request-payer requester` to all requests."""
 
+    @classmethod
+    async def from_config(cls, config: Config) -> S3Client:
+        """Creates an s3 client from a config.
+
+        Args:
+            config: The config object
+
+        Returns:
+            S3Client: A new s3 client
+        """
+        return cls(
+            requester_pays=config.s3_requester_pays, region_name=config.s3_region_name
+        )
+
     def __init__(
-        self, region_name: str = DEFAULT_REGION_NAME, requester_pays: bool = False
+        self,
+        requester_pays: bool = False,
+        region_name: str = DEFAULT_S3_REGION_NAME,
     ) -> None:
         super().__init__()
         self.session = aiobotocore.session.get_session()
         self.region_name = region_name
         self.requester_pays = requester_pays
 
-    async def open_url(self, url: URL) -> AsyncIterator[bytes]:
+    async def open_url(
+        self, url: URL, content_type: Optional[str] = None
+    ) -> AsyncIterator[bytes]:
         """Opens an s3 url and iterates over its bytes.
 
         Args:
             url: The url to open
+            content_type: The expected content type
 
         Yields:
             AsyncIterator[bytes]: An iterator over the file's bytes
@@ -49,12 +65,10 @@ class S3Client(Client):
         Raises:
             SchemeError: Raised if the url's scheme is not ``s3``
         """
-        if url.scheme != "s3":
-            raise SchemeError(f"only s3 urls are allowed: {url}")
         if self.requester_pays:
-            config = Config()
+            config = botocore.config.Config()
         else:
-            config = Config(signature_version=UNSIGNED)
+            config = botocore.config.Config(signature_version=UNSIGNED)
         async with self.session.create_client(
             "s3",
             region_name=self.region_name,
@@ -69,28 +83,13 @@ class S3Client(Client):
             if self.requester_pays:
                 params["RequestPayer"] = "requester"
             response = await client.get_object(**params)
+            print(response)
+            if content_type and response["ContentType"] != content_type:
+                raise ContentTypeError(
+                    actual=response["ContentType"], expected=content_type
+                )
             async for chunk in response["Body"]:
                 yield chunk
-
-    async def download_asset(self, key: str, asset: Asset, path: Path) -> None:
-        """Downloads an asset to a location.
-
-        If the initial download fails with a scheme error, the client looks for
-        an alternate href and tries again with that.
-
-        Args:
-            key: The asset key
-            asset: The asset
-            path: The destination path
-        """
-        try:
-            return await super().download_asset(key, asset, path)
-        except AssetDownloadError as err:
-            if isinstance(err.err, SchemeError):
-                maybe_asset = self._asset_with_alternate_href(asset)
-                if maybe_asset:
-                    return await super().download_asset(key, maybe_asset, path)
-            raise err
 
     async def has_credentials(self) -> bool:
         """Returns true if the sessions has credentials."""
@@ -105,24 +104,4 @@ class S3Client(Client):
         exc_val: Optional[BaseException],
         exc_tb: Optional[TracebackType],
     ) -> Optional[bool]:
-        return None
-
-    def _asset_with_alternate_href(self, asset: Asset) -> Optional[Asset]:
-        # TODO some of this logic could be refactored out to be more common, but
-        # not all (e.g. requester pays)
-        alternate = asset.extra_fields.get("alternate")
-        if alternate and isinstance(alternate, dict):
-            s3 = alternate.get("s3")
-            if s3 and isinstance(s3, dict):
-                requester_pays = s3.get("storage:requester_pays", False)
-                href = s3.get("href")
-                platform = s3.get("storage:platform", "AWS")
-                if platform == "AWS" and href:
-                    if requester_pays and not self.requester_pays:
-                        raise ValueError(
-                            f"alternate href {href} requires requester pays, but "
-                            "requester pays is not enabled on this client"
-                        )
-                    asset.href = href
-                    return asset
         return None