+++ tests/test_earthdata_client.py
@@ -16,6 +16,6 @@ pytestmark = [
 
 async def test_download_href(tmp_path: Path) -> None:
     href = "https://data.lpdaac.earthdatacloud.nasa.gov/lp-prod-protected/MYD11A1.061/MYD11A1.A2023145.h14v17.061.2023146183035/MYD11A1.A2023145.h14v17.061.2023146183035.hdf"
-    async with await EarthdataClient.default() as client:
+    async with await EarthdataClient.login() as client:
         await client.download_href(href, tmp_path / "out.hdf")
         assert os.path.getsize(tmp_path / "out.hdf") == 197419